refactor(tests): merged two folders tests > services
